move responsibility for destroying messages from LocalMessage to LocalFolder
